// src/types/Opportunity.ts
export interface Opportunity {
    id: string;
    name: string;
    amount: number;
    stageName: string;
    probability: number;
    description: string;
    createdDate: string;
    lastModifiedDate: string;
    accountId: string;
    accountName: string;
    ownerId: string;
    ownerName: string;
} 